+++ src/components/TeacherComponents/OnboardingForm/onboardingActions/getSubscriptionStatus.ts
@@ -1,8 +1,8 @@
 import OakError from "@/errors/OakError";
-import { subscriptionResponseSchema } from "@/pages/api/hubspot/subscription";
+import handleFetchError from "@/utils/handleFetchError";
+import { invariant } from "@/utils/invariant";
 
 export async function getSubscriptionStatus(
-  email: string,
   callback?: (status: boolean) => void,
 ) {
   try {
@@ -12,11 +12,15 @@ export async function getSubscriptionStatus(
         "Content-Type": "application/json",
       },
       body: JSON.stringify({
-        email,
         subscriptionName: "School Support",
       }),
     });
-    const result = subscriptionResponseSchema.parse(await response.json());
+    handleFetchError(response);
+
+    const result = await response.json();
+
+    invariant(typeof result === "boolean", "Expected response to be a boolean");
+
     if (callback) {
       callback(result);
     }